Remove cyclical module dependency
