+++ packages/liveblocks-core/src/crdts/UpdateDelta.ts
@@ -1,0 +1,7 @@
+export type UpdateDelta =
+  | {
+      type: "update";
+    }
+  | {
+      type: "delete";
+    };
